API

Inputs

UncertaintyQuantification.sample — Method

sample(inputs::Array{<:UQInput}, n::Integer)

Generates n correlated samples from a collection of inputs. Returns a DataFrame

See also: RandomVariable, Parameter

Parameter

UncertaintyQuantification.Parameter — Type

Parameter(value::Real, name::Symbol)

Defines a parameter value (scalar), with an input value and a name.

Examples

julia> Parameter(3.14, :π)
Parameter(3.14, :π)

RandomVariable

UncertaintyQuantification.RandomVariable — Type

RandomVariable(dist::Sampleable{Univariate}, name::Symbol)

Defines a random variable, with a univariate distribution from Distributions.jl and a name.

Examples

julia> RandomVariable(Normal(), :x)
RandomVariable(Normal{Float64}(μ=0.0, σ=1.0), :x)

julia> RandomVariable(Exponential(1), :x)
RandomVariable(Exponential{Float64}(θ=1.0), :x)

UncertaintyQuantification.sample — Method

sample(rv::RandomVariable, n::Integer)

Generates n samples from a random variable. Returns a DataFrame.

Examples

x = RandomVariable(Normal(), :x)
sample(x, 3)

# output

3×1 DataFrame
│ Row │ x          │
│     │ Float64    │
├─────┼────────────┤
│ 1   │ 1.00763    │
│ 2   │ -0.643618  │
│ 3   │ -0.0826132 │

See also: RandomVariable

Index

UncertaintyQuantification.Parameter

UncertaintyQuantification.RandomVariable

UncertaintyQuantification.sample

UncertaintyQuantification.sample