Bayesian Updating

Markov Chain Monte Carlo Methods for Bayesian updating.

Index

UncertaintyQuantification.AbstractBayesianMethod

UncertaintyQuantification.SingleComponentMetropolisHastings

UncertaintyQuantification.TransitionalMarkovChainMonteCarlo

UncertaintyQuantification.bayesianupdating

Types

UncertaintyQuantification.AbstractBayesianMethod — Type

AbstractBayesianMethod

Subtypes are used to dispatch to the differenct MCMC methods in bayesianupdating.

Subtypes are:

SingleComponentMetropolisHastings

TransitionalMarkovChainMonteCarlo

UncertaintyQuantification.SingleComponentMetropolisHastings — Type

SingleComponentMetropolisHastings(proposal, x0, n, burnin, islog)

Passed to bayesianupdating to run the single-component Metropolis-Hastings algorithm starting from x0 with univariate proposal distibution proposal. Will generate n samples after performing burnin steps of the Markov chain and discarding the samples. The flag islog specifies whether the prior and likelihood functions passed to the bayesianupdating method are already given as logarithms.

Alternative constructor

    SingleComponentMetropolisHastings(proposal, x0, n, burnin)  # `islog` = true

UncertaintyQuantification.TransitionalMarkovChainMonteCarlo — Type

TransitionalMarkovChainMonteCarlo(prior, n, burnin, β, islog)

Passed to [`bayesianupdating`](@ref) to run thetransitional Markov chain Monte Carlo algorithm  with [`RandomVariable'](@ref) vector `prior`. At each transitional level, one sample will be generated from `n` independent Markov chains after `burnin` steps have been discarded. The flag `islog` specifies whether the prior and likelihood functions passed to the  [`bayesianupdating`](@ref) method are already  given as logarithms.

Alternative constructors

    TransitionalMarkovChainMonteCarlo(prior, n, burnin, β)  # `islog` = true
     TransitionalMarkovChainMonteCarlo(prior, n, burnin)    # `β` = 0.2,  `islog` = true

References

[10]

Methods

UncertaintyQuantification.bayesianupdating — Function

bayesianupdating(prior, likelihood, models, mcmc)

Perform bayesian updating using the given prior, likelihood, models and any MCMC sampler AbstractBayesianMethod.

Alternatively the method can be called without models.

bayesianupdating(prior, likelihood, mcmc)

When using TransitionalMarkovChainMonteCarlo the prior can automatically be constructed.

bayesinupdating(likelihood, models, tmcmc)
bayesianupdating(likelihood, tmcmc)

Notes

likelihood is a Julia function which must be defined in terms of a DataFrame of samples, and must evaluate the likelihood for each row of the DataFrame

For example, a loglikelihood based on normal distribution using 'Data':

likelihood(df) = [sum(logpdf.(Normal.(df_i.x, 1), Data)) for df_i in eachrow(df)]

If a model evaluation is required to evaluate the likelihood, a vector of UQModels must be passed to bayesianupdating. For example if the variable x above is the output of a numerical model.