Models

Index

UncertaintyQuantification.Model

UncertaintyQuantification.ParallelModel

UncertaintyQuantification.UQModel

UncertaintyQuantification.evaluate!

UncertaintyQuantification.evaluate!

Types

UncertaintyQuantification.UQModel — Type

Abstract supertype for all model types

UncertaintyQuantification.Model — Type

Model(f::Function, name::Symbol)

The function f must accept a DataFrame and return the result of the model for each row in the DataFrame as a vector. The name is used to add the output to the DataFrame.

UncertaintyQuantification.ParallelModel — Type

ParallelModel(f::Function, name::Symbol)

The ParallelModel does what the Model does with a small difference. The function f is passed a DataFrameRow not the full DataFrame. If workers (through Distributed) are present, the rows are evaluated in parallel.

Methods

UncertaintyQuantification.evaluate! — Method

evaluate!(m::Model, df::DataFrame)

Calls m.func with df and adds the result to the DataFrame as a column m.name

UncertaintyQuantification.evaluate! — Method

evaluate!(m::Model, df::DataFrame)

Calls m.func for each row of df and adds the result to the DataFrame as a column m.name. If workers are added through Distributed, the rows will be evaluated in parallel.