Simulations

Various Monte Carlo based simulations for a wide range of applications.

Index

UncertaintyQuantification.SubSetInfinity

UncertaintyQuantification.SubSetSimulation

Types

UncertaintyQuantification.SubSetSimulation — Type

SubSetSimulation(n::Integer, target::Float64, levels::Integer, proposal::UnivariateDistribution)

Defines the properties of a Subset simulation where n is the number of initial samples, target is the target probability of failure at each level, levels is the maximum number of levels and proposal is the proposal distribution for the markov chain monte carlo.

Examples

julia> SubSetSimulation(100, 0.1, 10, Uniform(-0.2, 0.2))
SubSetSimulation(100, 0.1, 10, Uniform{Float64}(a=-0.2, b=0.2))

References

[4]

UncertaintyQuantification.SubSetInfinity — Type

SubSetInfinity(n::Integer, target::Float64, levels::Integer, s::Real)

Defines the properties of a Subset-∞ simulation where n is the number of initial samples, target is the target probability of failure at each level, levels is the maximum number of levels and s is the standard deviation for the proposal samples.

Examples

julia> SubSetInfinity(100, 0.1, 10, 0.5)
SubSetInfinity(100, 0.1, 10, 0.5)

References

[6]

[7]